+++ src/app/home/home.component.ts
@@ -13,5 +13,4 @@ export class HomeComponent implements OnInit {
 
   ngOnInit() {
   }
-
 }